Move signer to types folder
