+++ libs/coin-evm/src/__tests__/unit/signOperation.unit.test.ts
@@ -3,8 +3,8 @@ import { Account } from "@ledgerhq/types-live";
 import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 import { SignerContext } from "@ledgerhq/coin-framework/signer";
 import { getCryptoCurrencyById, listCryptoCurrencies } from "@ledgerhq/cryptoassets";
+import type { EvmAddress, EvmSignature, EvmSigner } from "../../types/signer";
 import { buildSignOperation, applyEIP155 } from "../../signOperation";
-import { EvmAddress, EvmSignature, EvmSigner } from "../../signer";
 import { Transaction as EvmTransaction } from "../../types";
 import { makeAccount } from "../fixtures/common.fixtures";
 import * as nodeApi from "../../api/node/rpc.common";
